Make bag sharable via link
When creating a user, a `bag_code` is generated. When opening
`/bag/:bag_code`, the user to which the `bag_code` belongs is logged
in, and redirected to `/app/bag`.

The link to the bag is accessible via the bag modal for a regularly
logged in user, so whoever starts the game can share it with their team.